// types/enums.ts
export enum Season {
  Spring = 'Spring',
  Summer = 'Summer',
  Fall = 'Fall',
  Winter = 'Winter',
}

export enum Weather {
  Sunny = 'Sunny',
  Rainy = 'Rainy',
  Cloudy = 'Cloudy',
  Snowy = 'Snowy',
  Hot = 'Hot',
  Cold = 'Cold',
  Humid = 'Humid',
  Dry = 'Dry',
}

export enum TimeOfDay {
  Morning = 'Morning',
  Afternoon = 'Afternoon',
  Evening = 'Evening',
  Night = 'Night',
}

export enum NoteFamily {
  Citrus = 'Citrus',
  Floral = 'Floral',
  Woody = 'Woody',
  Oriental = 'Oriental',
  Fresh = 'Fresh',
  Spicy = 'Spicy',
  Green = 'Green',
  Aquatic = 'Aquatic',
}